Feature/#77 감정일기 생성 시 이미지필드 없어도 저장 가능하게 로직 변경

+++ src/main/java/com/example/mutsideout_mju/service/DiaryService.java
@@ -17,6 +17,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartException;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -73,7 +74,13 @@ public class DiaryService {
                 .build();
         diary = diaryRepository.save(diary);
 
-        imageService.uploadImages(user, diary, images);
+        try {
+            if (images != null && !images.isEmpty()) {
+                imageService.uploadImages(user, diary, images);
+            }
+        } catch (MultipartException e) {
+            throw new MultipartException(e.getMessage());
+        }
     }
 
     /**
@@ -82,10 +89,16 @@ public class DiaryService {
     @Transactional
     public void updateDiaryById(User user, UUID diaryId, UpdateDiaryDto updateDiaryDto, List<MultipartFile> images) throws IOException {
         Diary newDiary = findDiary(user.getId(), diaryId);
-        if (images != null) {
-            imageService.deleteImages(newDiary);
-            imageService.uploadImages(user, newDiary, images);
+
+        try {
+            if (images != null && !images.isEmpty()) {
+                imageService.deleteImages(newDiary);
+                imageService.uploadImages(user, newDiary, images);
+            }
+        } catch (MultipartException e) {
+            throw new MultipartException(e.getMessage());
         }
+
         newDiary.setTitle(updateDiaryDto.getTitle())
                 .setContent(updateDiaryDto.getContent());
